Major progress, Fixed Pause and Resume, Config work started
Fixed the Pause and Resume, Daemon pointed out need to just SetSynchDirty so the client would be updated to.

Entered Config entries for Requiring Battery
or
Not Requiring Battery
For Processing

Introduced Conditions for checking for the config for Require Battery or not

Introduced Battery Usage per operation ad defined in the config

Introduced Plastic Wrap Usage Per Operation as defined in the config

Started working on config rework but dont think its really ready for it.

Tomarrow will work more on the config settings and such, then talking to Iceblade more on anything missed or such. Hopefully Testing this week.