Second Attempt at #54
Finally!!!!

So this was overall tricky but let me explain what I did by starting with what we couldnt do

So first we cant Stop and delete a timer without crashing the server in the same function, however triggering a Stop timer function and a delete timer function one right after the other is took quick talking miliseconds to quick.

So the Solution was the create a New timer that is only triggered when the Table Ran out of materials.

This timer was triggered by the Check in Do_processing when it ran out of mats, however the else statement in that void would trigger then it would trigger the new timer to start with a 1 second defined in the new timer void call.

so the New timers void was called  End_Processing(#) # = timer before timer ran its function tied to it. new timer = m_CP_ProcessingCheck

This then would trigger the StopProduction() and stop the m_CP_ProcessingCheck. If a player started Do_Processing again once more materials were attached it would then run this system all over again

TLDR : I made a Timer to space the time between actions to stop causing the server crash's when the timer was deleted faster then the server could register the timer was stopped.