Fixed Null Pointers and failed to spawn entity issues
Issues was Fault of  incorrectly position checking and results being empty

so first off the Result being ("") or rather empty was the reason for error `failed to spawn entity issues`

Once the Results area or lines 49 to 58 for example were commented out that error was resolved without affecting the recipes operation and outcome

Second was the Null Pointers to line 84.

This was cause two things were happening that i failed to see before

1. Variable Budsholder is just a place holder and is attached to nothing but itembase.
so when fetching the GetPosition () in line 84 it didnt know where the position was as BudsHolder didnt exist in the game as item so couldnt find its postion.

2.  BudsHolder in line 84 was being set too Itembase and then casted, Item base doesnt have the Config.cpp Value GetBag() Line 81 checks for and assigns to Bagname. (honestly this should of broke the recipe but it didnt)

Code wise how I fixed it

So the result was to comment out the Itembase BudsHolder line 78.

Then Make line 84 state the Class first, then Var name it will use, Then Cast the class to the item creation and finally have it Check ingredient1 position and spawn the created item there instead as ingredient1 exists.